requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
configparser==7.2.0
extensions==0.4
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.3.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
pymysql==1.1.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
email-validator==2.1.0.post1
